Reworked so that the field *always* displays a valid value.
Previously invalid values were allowed because the user could be on their
way to typing valid value. The field installed itself as a FocusListener
and checked the value when focus was lost. This wasn't ideal, as a user
could enter what they thought was a valid value and then click a button and
the action of them clicking the button would instantly change the displayed
value with no chance for them to see that it changed from they had last
entered.

Now it does clever highlighting to let you edit the field while always
keeping the value between the minimum and the maximum.


git-svn-id: https://samskivert.googlecode.com/svn/trunk@1583 6335cc39-0255-0410-8fd6-9bcaacd3b74c