Altered so that document events are never fired while the document contains an invalid value. This was made extra hairy by the fact that the FilterBypass installed on a Document implements replace() by simply calling remove() and then insert(). That's fine and dandy and saves lots of code, I'm sure, but the each of them fires an event, resulting in two events when a replace() happens.
I Created a special Document that suppresses 'remove' events from firing.
Not a problem, because I've also structured the filter such that removes()
end up being implemented as a replace() (due to the filtering and
processing that we do to ensure that valid, properly formatted integers
are always displayed) so the 'insert' event will take care of all our needs.


git-svn-id: https://samskivert.googlecode.com/svn/trunk@1585 6335cc39-0255-0410-8fd6-9bcaacd3b74c