I grow tired of having a commons-io dependency (and all the crap that comes along with it just to have essentially two methods:
copy(InputStream,OutputStream) and toString(InputStream)

We'll just duplicate those ten lines of code and save a dozen of our libraries
some trouble. This is a precursor to a possible Velocity jihad because Velocity
is pissing me off by not working because of some bullshit classloader
tomfoolery that is the sort of thing that makes Java programmers want to jump
butt first off the building.


git-svn-id: https://samskivert.googlecode.com/svn/trunk@2586 6335cc39-0255-0410-8fd6-9bcaacd3b74c