Ok, Charlie Groves has done a good job convincing me I'm crazy, but I'll check this in anyway, at least for posterity.
The advantages over the 'CountingMap' (currently in yohoho):
- Faster, because adding to a count doesn't add to a boxed int, nor does
  it hash twice (except on the intial use of a particular key).

The disadvantages:
- More complicated implementation (CountingMap builds upon google
  collections)
- If a non-HashMap is desired, the fact that the underlying map is
  <K, int[]> is exposed.
- When iterating over the entrySet, each entry is a new object
  created solely for your iteration.

So.. yeah. I'll check it in. Maybe I'll blow it away. We should
probably think about how samskivert will move forward- do we want to
depend on google-collect or guava? Do we want to deprecate huge
swaths of this library in preference to those? A lot of what's
in here is a bit outdated, but maintained to support code built
atop it.


git-svn-id: https://samskivert.googlecode.com/svn/trunk@2704 6335cc39-0255-0410-8fd6-9bcaacd3b74c