1.7 javac is apparently more perceptive than ejc when it comes to warning about raw type usage. I'm surprised that the Eclipse code hygienist failed to catch these. Also had to be slightly more blatant in a few places where we're subverting the generics type system because the 1.7 javac is less tolerant of the fast and loose.
For future reference, that usually means first assigning a reference to a
wildcard and then casting that to what you want:

Foo<?> fc = someFooWithCrazyAssParameters;
@SuppressWarnings("unchecked") Foo<Object> fo = (Foo<Object>)fc;

instead of just:

@SuppressWarnings("unchecked") Foo<Object> fo =
    (Foo<Object>)someFooWithCrazyAssParameters;

because javac won't let you do casts that are inconvertible and it's smarter
about what sort of generic types are convertible to one another.

Hopefully this future reference won't be needed because we won't be needing to
subvert the type system. But when reflection is involved, it's hard not to end
up doing a bit of the old Father Knows Best.


git-svn-id: https://samskivert.googlecode.com/svn/trunk@2756 6335cc39-0255-0410-8fd6-9bcaacd3b74c