Order our equals() comparisons in the same way Java Collections do.
ArrayList and friends explicitly call equals() on the supplied object with the
array object as the argument, rather than the other way around. This allows one
to, for example, pass in some sort of Key object that can compare itself for
equality against a list of KeyedValue objects.

This sort of delicate dance around the equality contract is dangerous, but we
should let the caller decide whether they wish to dance with the devil.