Get rid of TestUtil and related hackery.
The SiteResourceLoader is some ancient stuff that's not used anywhere any more,
and there's no use in keeping this crufty crap around to test it. I'd nix it
too but I don't want to cause yet more trouble.

TestUtil was weird and a bad idea and good riddance to it as well.